% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airway.R
\docType{data}
\name{airway}
\alias{airway}
\title{Airway smooth muscle cells}
\format{
A data frame with 63677 rows and 8 columns. Each row corresponds to a
specific gene and each column to a labeled sample.
}
\source{
FASTQ files from SRA, phenotypic data from GEO
}
\usage{
data("airway")
}
\description{
Read counts per gene for airway smooth muscle cell lines RNA-Seq experiment
}
\details{
The cell lines of the even-numbered samples were treated with dexamethasone,
whereas the cell lines of the odd-numbered samples were not.
}
\references{
Himes, B. E., Jiang, X., Wagner, P., Hu, R., Wang, Q., Klanderman, B.,
   Whitaker, R. M., Duan, Q., Lasky-Su, J., Nikolos, C., Jester, W.,
   Johnson, M., Panettieri, R. Jr., Tantisira, K. G., Weiss, S. T., Lu, Q.
   (2014). RNA-Seq Transcriptome Profiling Identifies CRISPLD2 as a
   Glucocorticoid Responsive Gene that Modulates Cytokine Function in Airway
   Smooth Muscle Cells. \emph{PLoS One} \strong{9}(6).
   \doi{https://doi.org/10.1371/journal.pone.0099625}
}
\keyword{datasets}
