% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteDatasets.R
\docType{package}
\name{DiscreteDatasets-package}
\alias{DiscreteDatasets}
\alias{DiscreteDatasets-package}
\title{DiscreteDatasets}
\description{
This package contains example datasets for use with discrete statistical
tests and discrete multiple testing procedures. Some of them are also
available as a four-column version, so that each row represents a 2x2 table.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DISOhda/DiscreteDatasets}
  \item Report bugs at \url{https://github.com/DISOhda/DiscreteDatasets/issues}
}

}
\author{
\strong{Maintainer}: Florian Junge \email{diso.fbmn@h-da.de}

Authors:
\itemize{
  \item Christina Kihn
  \item Sebastian Döhler
}

Other contributors:
\itemize{
  \item Lukas Klein [contributor]
}

}
