% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteDatasets.R
\docType{package}
\name{DiscreteDatasets}
\alias{DiscreteDatasets}
\alias{DiscreteDatasets-package}
\title{DiscreteDatasets}
\description{
This package contains example datasets for use with discrete statistical
tests and discrete multiple testing procedures.
}
