% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federalist.R
\docType{data}
\name{federalist}
\alias{federalist}
\title{The Federalist Papers word count data}
\format{
\code{federalist} is a \code{data.frame} with 77 rows and 1984
columns:
\describe{
  \item{\code{doc_no}}{Article number}
  \item{\code{doc_author}}{Author of the article}
  \item{\code{...}}{The remaining 1982 columns are the word counts}
}
}
\usage{
data("federalist")
}
\description{
Author assignments and counts of the 1500 most common
words from "The Federalist" articles for Madison and Hamilton.

"The Federalist Papers" are a set of 85 articles written under the pseudonym
"Publius" to promote the ratification of the US constitution by Alexander
Hamilton, James Madison and John Jay in 1787 and 1788. There are multiple
sources which attribute the articles to their real
authors. This author attribution has been a popular problem in natural
language processing.

The data was prepared with the following steps by employing the
\code{tm} package:
 \enumerate{
   \item{Load the texts from the \code{syllogi} package,}
   \item{Lowercase,}
   \item{Remove punctuation,}
   \item{Strip whitespace,}
   \item{Drop texts from Jay, Madison+Hamilton together, and the
         redundant version of article 70,}
   \item{Find the 1500 most common words for each author,}
   \item{Count the occurrences of these words in the texts}
}
}
\references{
Donoho, D. L., & Kipnis, A. (2022). Higher criticism to compare two large
   frequency tables, with sensitivity to possible rare and weak differences.
   \emph{Annals of Statistics}, \strong{50}(3), 1447-1472.
   \doi{10.1214/21-AOS2158}

Feinerer, I., & Hornik, K. (2024). tm: Text Mining Package. R package
   version 0.7-15. \emph{CRAN}. \url{https://CRAN.R-project.org/package=tm}

Studyvin, J. (2024). syllogi: Collection of Data Sets for Teaching Purposes.
   R package version 1.0.3. \emph{CRAN}.
   \url{https://CRAN.R-project.org/package=syllogi}
}
\keyword{datasets}
