% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteDatasets.R
\docType{package}
\name{DiscreteDatasets-package}
\alias{DiscreteDatasets}
\alias{DiscreteDatasets-package}
\title{DiscreteDatasets}
\description{
This package contains example datasets for use with discrete statistical
tests and discrete multiple testing procedures. Some of them are also
available as a four-column version, so that each row represents a 2x2 table.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DISOhda/DiscreteDatasets}
  \item Report bugs at \url{https://github.com/DISOhda/DiscreteDatasets/issues}
}

}
\author{
\strong{Maintainer}: Florian Junge \email{diso.fbmn@h-da.de} (\href{https://orcid.org/0009-0001-6856-6938}{ORCID})

Authors:
\itemize{
  \item Christina Kihn
  \item Sebastian Döhler \email{sebastian.doehler@h-da.de} (\href{https://orcid.org/0000-0002-0321-6355}{ORCID})
}

Other contributors:
\itemize{
  \item Lukas Klein (\href{https://orcid.org/0000-0002-1413-497X}{ORCID}) [contributor]
  \item Iqraa Meah [contributor]
}

}
